refactor: extract preset item models and clean up dead code
